Filtering, rearrangement and scanning detection
The code efficiently filters the data and rearranges according to the detected signal in each frame and at the end detects the trajectories. 
There are 5 file as example of raw data to check the code. 
The inputs of this code are file-paths and visual x,y limits and the output is the trajectories. 
The code might need substantial technical revision from writing style and object names to plotting configurations!